requirements.txt
Flask==1.1.2
tensorflow==1.14.0
matplotlib==3.2.2
numpy==1.18.5
pillow==7.1.2
WTForms==2.2.1
Flask_WTF==0.14.3
Werkzeug==0.16.0